Added a failsafe from my unpushed master branch becouse im dumb and forgot about changing the branch also deleted unused files
